refactor(mcp-visualizer): extract McpVisualizerOrchestra factory
Move the Orchestra-with-visualizer-callbacks wiring out of RunTool into
its own file so the run path stays focused on flow execution.
